Add ability for multiple owners, Chase 5/24 overhaul

+++ app/src/main/java/com/davidlcassidy/travelwallet/Database/MainDatabaseHelper.java
@@ -23,7 +23,7 @@ public class MainDatabaseHelper extends SQLiteOpenHelper {
     // date values in local database unreadable and make the app unstable. If new fields are
     // required, consider using futureText and futureInteger fields rather than making structural
     // changes to the database.
-    public static final int DATABASE_VERSION = 1;
+    public static final int DATABASE_VERSION = 2;
     public static final SimpleDateFormat DATABASE_DATEFORMAT = new SimpleDateFormat("yyyy-MM-dd HH:mm:ss");
 
     private static final String DATABASE_NAME = "DatabaseMain.db";
@@ -59,7 +59,7 @@ public class MainDatabaseHelper extends SQLiteOpenHelper {
     public static final String COLUMN_CC_CLOSEDATE = "closeDate";
     public static final String COLUMN_CC_NOTIFICATIONSTATUS = "notificationStatus";
     public static final String COLUMN_CC_NOTES = "notes";
-    public static final String COLUMN_CC_FUTURETEXT1 = "futureText1";
+    public static final String COLUMN_CC_CREDITLIMIT = "futureText1";
     public static final String COLUMN_CC_FUTURETEXT2 = "futureText2";
     public static final String COLUMN_CC_FUTURETEXT3 = "futureText3";
     public static final String COLUMN_CC_FUTURETEXT4 = "futureText4";
@@ -69,6 +69,21 @@ public class MainDatabaseHelper extends SQLiteOpenHelper {
     public static final String COLUMN_CC_FUTUREINTEGER3 = "futureInteger3";
     public static final String COLUMN_CC_FUTUREINTEGER4 = "futureInteger4";
     public static final String COLUMN_CC_FUTUREINTEGER5 = "futureInteger5";
+
+    public static final String TABLE_O = "users";
+    public static final String COLUMN_O_ID = "_id";
+    public static final String COLUMN_O_NAME = "name";
+    public static final String COLUMN_O_NOTES = "notes";
+    public static final String COLUMN_O_FUTURETEXT1 = "futureText1";
+    public static final String COLUMN_O_FUTURETEXT2 = "futureText2";
+    public static final String COLUMN_O_FUTURETEXT3 = "futureText3";
+    public static final String COLUMN_O_FUTURETEXT4 = "futureText4";
+    public static final String COLUMN_O_FUTURETEXT5 = "futureText5";
+    public static final String COLUMN_O_FUTUREINTEGER1 = "futureInteger1";
+    public static final String COLUMN_O_FUTUREINTEGER2 = "futureInteger2";
+    public static final String COLUMN_O_FUTUREINTEGER3 = "futureInteger3";
+    public static final String COLUMN_O_FUTUREINTEGER4 = "futureInteger4";
+    public static final String COLUMN_O_FUTUREINTEGER5 = "futureInteger5";
 
     private static final String TABLE_LOYALTYPROGRAMS_CREATE = "create table IF NOT EXISTS "
             + TABLE_LP + "( "
@@ -104,7 +119,7 @@ public class MainDatabaseHelper extends SQLiteOpenHelper {
             + COLUMN_CC_CLOSEDATE + " text, "
             + COLUMN_CC_NOTIFICATIONSTATUS + " integer, "
             + COLUMN_CC_NOTES + " text, "
-            + COLUMN_CC_FUTURETEXT1 + " text, "
+            + COLUMN_CC_CREDITLIMIT + " text, "
             + COLUMN_CC_FUTURETEXT2 + " text, "
             + COLUMN_CC_FUTURETEXT3 + " text, "
             + COLUMN_CC_FUTURETEXT4 + " text, "
@@ -114,6 +129,23 @@ public class MainDatabaseHelper extends SQLiteOpenHelper {
             + COLUMN_CC_FUTUREINTEGER3 + " integer, "
             + COLUMN_CC_FUTUREINTEGER4 + " integer, "
             + COLUMN_CC_FUTUREINTEGER5 + " integer "
+            +");";
+
+    private static final String TABLE_OWNERS_CREATE = "create table IF NOT EXISTS "
+            + TABLE_O + "( "
+            + COLUMN_O_ID + " integer primary key autoincrement, "
+            + COLUMN_O_NAME + " text, "
+            + COLUMN_O_NOTES + " text, "
+            + COLUMN_O_FUTURETEXT1 + " text, "
+            + COLUMN_O_FUTURETEXT2 + " text, "
+            + COLUMN_O_FUTURETEXT3 + " text, "
+            + COLUMN_O_FUTURETEXT4 + " text, "
+            + COLUMN_O_FUTURETEXT5 + " text, "
+            + COLUMN_O_FUTUREINTEGER1 + " integer, "
+            + COLUMN_O_FUTUREINTEGER2 + " integer, "
+            + COLUMN_O_FUTUREINTEGER3 + " integer, "
+            + COLUMN_O_FUTUREINTEGER4 + " integer, "
+            + COLUMN_O_FUTUREINTEGER5 + " integer "
             +");";
 
     public MainDatabaseHelper(Context context) {
@@ -128,13 +160,19 @@ public class MainDatabaseHelper extends SQLiteOpenHelper {
     public void onCreate(SQLiteDatabase database) {
         database.execSQL(TABLE_LOYALTYPROGRAMS_CREATE);
         database.execSQL(TABLE_CREDITCARDS_CREATE);
+        database.execSQL(TABLE_OWNERS_CREATE);
     }
 
     @Override
     public void onUpgrade(SQLiteDatabase db, int oldVersion, int newVersion) {
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_LP);
-        db.execSQL("DROP TABLE IF EXISTS " + TABLE_CC);
-        onCreate(db);
+        if (oldVersion == 1 && newVersion == 2){
+            db.execSQL(TABLE_OWNERS_CREATE);
+        } else {
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_LP);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_CC);
+            db.execSQL("DROP TABLE IF EXISTS " + TABLE_O);
+            onCreate(db);
+        }
     }
 
 }
